ARC-3497: align FE with backend
